common: malloc: cheri: Change pointer casting to uintptr_t
Change pointer casting from ulong to uintptr_t to match the
pointer type and cast the pointer to ulong for extracting the address
for the architectures the address at the lower part of the pointer type.
(e.g. CHERI).

+++ arch/arm/mach-k3/am65x/am654_init.c
@@ -208,7 +208,7 @@ void board_init_f(ulong dummy)
 	if (!pool_addr)
 		panic("ERROR: Can't allocate full malloc pool!\n");
 
-	mem_malloc_init((ulong)pool_addr, (ulong)pool_size);
+	mem_malloc_init((uintptr_t)pool_addr, (ulong)pool_size);
 	gd->flags |= GD_FLG_FULL_MALLOC_INIT;
 	debug("%s: initialized an early full malloc pool at 0x%08lx of 0x%lx bytes\n",
 	      __func__, (unsigned long)pool_addr, (unsigned long)pool_size);
